TODO: Change aco

+++ aa_6/parameters.h
@@ -7,7 +7,7 @@ typedef struct
 	double beta;
 	double Q;
 	double rho;
-	double time;
+	int time;
 } parameters_t;
 
 #endif
